Relax -Wmissing-field-initializers; fix remaining -Werror breaks
The Linux GCC build (-Werror) has apparently never run clean on
master: a local -k0 build surfaced ~100 missing-field-initializer
errors (test/options structs gained fields; call sites deliberately
only set the relevant ones - idiomatic partial aggregate-init, not a
bug) plus a handful of real small issues, fixed here:

- Wno-missing-field-initializers added to the shared GCC/Clang
  warning flags (MSVC unaffected; it doesn't warn on this).
- irem_m78: drop unused expand_3bit (same dead-helper pattern as the
  earlier irem_m52::in_range fix).
- irem_m82 adapter/system tests: narrowing ternary (unsigned int ->
  uint8_t) in identified_m82_source rank init, fixed with an explicit
  cast.
- amiga_adapter: drop unused old_state local (old_ports, which it was
  sliced from, is still used below for keyboard-edge routing).
- Full pinned-clang-format pass on every file touched here, since
  each carried pre-existing line-wrap drift unrelated to these fixes.

+++ src/apps/player/adapters/irem_m82/tests/irem_m82_adapter_test.cpp
@@ -158,8 +158,9 @@ namespace {
             if (!known.contains(set_id)) {
                 return std::nullopt;
             }
-            return identified_m82_source{.set_id = set_id,
-                                         .rank = set_id == "dkgensanm82" ? 0U : 2U};
+            return identified_m82_source{
+                .set_id = set_id,
+                .rank = static_cast<std::uint8_t>(set_id == "dkgensanm82" ? 0U : 2U)};
         }
         if (!std::filesystem::is_regular_file(path, ec) ||
             !ends_with_zip(path.filename().string())) {
@@ -316,11 +317,8 @@ TEST_CASE("irem_m82_adapter validates real M82 ROM sets", "[irem_m82][data]") {
     const auto sources = index_m82_source_roots(roots);
     const auto expected_sets = embedded_set_names();
     const std::map<std::string, std::string, std::less<>> parents{
-        {"airduelu", "airduel"},
-        {"majtitlej", "majtitle"},
-        {"rtype2j", "rtype2"},
-        {"rtype2jc", "rtype2"},
-        {"rtype2m82b", "rtype2"},
+        {"airduelu", "airduel"}, {"majtitlej", "majtitle"}, {"rtype2j", "rtype2"},
+        {"rtype2jc", "rtype2"},  {"rtype2m82b", "rtype2"},
     };
     for (const auto& set_name : expected_sets) {
         INFO("set=" << set_name);
